requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-login==1.9.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
outcome==1.2.0
passlib==1.7.4
psycopg2-binary==2.9.7
pydantic==2.4.0
pydantic_core==2.10.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
starlette==0.27.0
trio==0.21.0
typing_extensions==4.8.0
uvicorn==0.23.2
